+++ bmak-ecommerce.Domain/Entities/Catalog/ProductStock.cs
@@ -11,7 +11,7 @@ namespace bmak_ecommerce.Domain.Entities.Catalog
     {
         public string WarehouseName { get; set; }
         public string BatchNumber { get; set; } // Số lô (Quan trọng cho gạch)
-        public float QuantityOnHand { get; set; }
+        public int QuantityOnHand { get; set; }
 
         public int ProductId { get; set; }
         public virtual Product Product { get; set; }
